+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 6.11.0 (March 1, 2019)
 ### New Features
+- [#898](https://github.com/openscope/openscope/issues/898) - Add command to exit holding pattern
 
 
 ### Bugfixes